flask-task-api/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-CORS==4.0.0
flasgger==0.9.7.1
python-dotenv==1.0.0
SQLAlchemy==1.4.49
marshmallow==3.20.1
bcrypt==4.0.1
gunicorn==21.2.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0
python-json-logger==2.0.7
redis==5.0.1
